// widgets/settings/_shared/types.ts
export type FormVType = {
	firstName?: string | undefined;
	title?: string | undefined;
	email?: string | undefined;
	bio?: string | undefined;
	location?: string | undefined;
	country?: string | undefined;
	tags?: string[] | undefined;
	isPrivate?: boolean | undefined;
	website?: string | undefined;
	x?: string | undefined;
	tiktok?: string | undefined;
	instagram?: string | undefined;
	github?: string | undefined;
};